Change modify_catalogs.py to take catalog paths on command line
Previously the script worked by taking an input directory with the
`-i` option (default: `input_catalogs`) and positional arguments
were interpreted as *basenames* of catalogs in this directory.

Now the positional arguments are interpreted as *paths* to the input
catalogs, and the `-i` option is removed. Modified catalogs are still
written to the output directory (see `-o`) with filename taken from
the basename of the input path.