Loop until cert has > 1 hour remaining before each publication step
Before each long-running publisher command, ensure there is at least one
hour remaining for the proxy certificate. If not, loop until there is.
This means that a batch-run will never exit early due to proxy expiry.

Also increase the required validity remaining at the start of the script
to 70 hours.