Only produce one aggregation per catalog in modify_catalogs.py
It turns out that there are actually no datasets for which we should have
multiple aggregations. Therefore always try to aggregate ALL netcdf files
within a catalog.

However it is still useful to keep partition_files.py so that we can print a
warning if the files in a catalog look hetrogeneous, so that this can be
further investigated.

To this end I also add a script `find_netcdf.py` that lists all netcdf files
referenced in a catalog; this can be used with the (modified)
`partition_files.py` script to see in what way the file paths in the catalog
differ.